services/prediction_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv
pydantic
scikit-learn
mlflow==2.17.2
pytest
httpx
requests